refactor: 유틸 함수 모으기
